代码结构调整
